add rewards chart to account details
