+++ data/rewards.js
@@ -148,3 +148,37 @@ export const useValidatorRewards = (
     isError: error,
   }
 }
+
+export const getAccountRewardsBuckets = async (
+  address,
+  numBack,
+  bucketType,
+) => {
+  const list = await client.account(address).rewards.sum.list({
+    minTime: `-${numBack} ${bucketType}`,
+    maxTime: new Date(),
+    bucket: bucketType,
+  })
+  const rewards = await list.take(TAKE_MAX)
+  return rewards.reverse()
+}
+
+export const useAccountRewards = (
+  address,
+  numBack = 30,
+  bucketType = 'day',
+) => {
+  const key = `rewards/accounts/${address}/${numBack}/${bucketType}`
+  const fetcher = (address, numBack, bucketType) => () =>
+    getAccountRewardsBuckets(address, numBack, bucketType)
+
+  const { data, error } = useSWR(key, fetcher(address, numBack, bucketType), {
+    refreshInterval: 1000 * 60 * 10,
+  })
+
+  return {
+    rewards: data,
+    isLoading: !error && !data,
+    isError: error,
+  }
+}
